summary class to import into run.py for inference
This script will load the T5 summary model in and perform inference given a string argument. This is a modification of run_local_RAC.py

Currently the summary model isn't good enough to perform much at all. I believe this has to with using rouge as the training metric. I am not sure the exact reason though, perhaps just not trained long enough or poor data. 

Am looking into it.