Implemented summary model --> bot
The T5 summary model I fine tuned did not reach a high enough level of performance to work within the bot. The changes made in this commit implemented the base T5 model for summary into the bot as a placeholder until I can get the fine tuned version implemented. I am also considering whether summarization is the appropriate task for my objective, which is to politely/cleanly explain why a message is innapropriate. After implementing the base T5 I am starting to think that multi class classification might be better and more informative to the violatng user. Although I do still want to pursue fine tuning some type model to achieve this clean summary idea. It just seems like the intended task is not the same as summarization, but rather it might be more in line with question answering. I'm looking into solutions.